refactor(comply): relocate single-caller helpers into their partitions
Four check_codegen helpers had exactly one caller; moving each into its
single-caller partition localizes complexity and drops the main file
from 393 to 284 lines.

  clause_ids_from_json     → check_codegen_attribute.rs
  codegen_receipt_outcome  → check_codegen_receipts.rs
  compile_profile_outcome  → check_codegen_compile_profiles.rs
  manifest_entries         → check_codegen_manifest_sha.rs

ReceiptOutcome enum stays in main (shared by receipts + compile_profiles).
After relocation, check_codegen.rs max cyclomatic drops 13→6 and max
cognitive drops 31→9.

Refs pmat-spec.md#work-contract-extensions-components-27-31
